Remove processing and postprocessing
