--- recOrder/postproc/__init__.py
@@ -1,1 +1,0 @@
-from .post_processing import *
